backend/ai_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.3.0+cpu
torchvision==0.18.0+cpu
Pillow==10.3.0

# For GPU servers replace the two lines above with:
# torch==2.3.0
# torchvision==0.18.0
